+++ frontend/app/post/_components/commentList.tsx
@@ -1,15 +1,5 @@
 import { mockComments } from "@/constants/mock";
-
-function createDateString(baseDate: Date): string {
-  const year: string = String(baseDate.getFullYear());
-  const month: string = String(baseDate.getMonth());
-  const day: string = String(baseDate.getDay());
-  const timeH: string = String(baseDate.getHours());
-  const timeM: string = String(baseDate.getMinutes());
-  const timeS: string = String(baseDate.getSeconds());
-  return year + "-" + month + "-" + day + " " + timeH + ":" + timeM + ":" + timeS;
-}
-
+import { createDateString } from "@/utils/index";
 const CommentList = () => {
   //const commentList = backend.comment.getCommentOld();
   return (
